refactor(notification): move scheduler services to scheduler package
Renamed and relocated NotificationSchedulerService and StreakResetSchedulerService to the scheduler package, updating their class names to NotificationScheduler and StreakResetScheduler. This improves project structure by grouping scheduler-related classes together.
